Moved searchBar into newly created components directory.
