+++ src/index.css
@@ -3,3 +3,9 @@ body {
   padding: 0;
   font-family: sans-serif;
 }
+
+
+.search-input {
+	width: 25%;
+    /* height: 5% */
+}
